+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adding group annotation content type
 - Add rejectDuplicates setting to workspace
 - Add `hideToolbar` to resume & JD search config
+- Add ExtractorConfig object to Collection
 
 ## [4.2.0] - 2023-03-20
 ### Fixed